+++ mentor_worker_benchmark/__init__.py
@@ -1,4 +1,4 @@
 """Mentor Worker Benchmark package."""
 
 __all__ = ["__version__"]
-__version__ = "1.2.2"
+__version__ = "1.2.3"